data_scraper/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pandas
requests
tqdm
python-dotenv
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.3.0
